케어테이커 인증요청 완료

+++ src/main/java/com/sopt/rescat/vo/JwtTokenVO.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class JwtTokenVO {
-    private int idx = -1;
+    private Long idx = (long) -1;
 }
